+++ src/test/renderApp.tsx
@@ -1,5 +1,6 @@
 import { render } from "@testing-library/react";
 import App from "../App";
+import { BoardsProvider } from "../boards/BoardsProvider";
 import { BoardProvider } from "../board/BoardProvider";
 import { ClipboardProvider } from "../board/ClipboardProvider";
 import { ThemeProvider } from "../theme/ThemeProvider";
@@ -7,11 +8,13 @@ import { ThemeProvider } from "../theme/ThemeProvider";
 export function renderApp() {
   return render(
     <ThemeProvider>
-      <BoardProvider>
-        <ClipboardProvider>
-          <App />
-        </ClipboardProvider>
-      </BoardProvider>
+      <BoardsProvider>
+        <BoardProvider>
+          <ClipboardProvider>
+            <App />
+          </ClipboardProvider>
+        </BoardProvider>
+      </BoardsProvider>
     </ThemeProvider>,
   );
 }